cleanup, move files and reexports
