+++ apps/directorate-of-equality-api/src/modules/report-comment/report-comment.controller.ts
@@ -14,7 +14,7 @@ import { ApiBearerAuth, ApiOperation, ApiResponse } from '@nestjs/swagger'
 import { TokenJwtAuthGuard } from '@dmr.is/shared-modules'
 
 import { CurrentReportResourceContext } from '../report/decorators/current-report-resource-context.decorator'
-import { ReportResourceGuard } from '../report/guards/report-resource.guard'
+import { ReportResourceGuard } from '../../core/guards/report-resource/report-resource.guard'
 import { CreateReportCommentDto } from './dto/create-report-comment.dto'
 import { ReportCommentDto } from './dto/report-comment.dto'
 import { type ReportResourceContext } from '../report/types/report-resource-context'
