--- apps/directorate-of-equality-api/src/modules/report-comment/types/report-comment-context.ts
@@ -1,4 +1,0 @@
-export enum CommentAuthorKindEnum {
-  REVIEWER = 'REVIEWER',
-  COMPANY_ADMIN = 'COMPANY_ADMIN',
-}
