--- apps/directorate-of-equality-api/src/modules/report/dto/report-comment.dto.ts
@@ -1,1 +1,0 @@
-export { ReportCommentDto } from '../../report-comment/dto/report-comment.dto'
